adding finance_calculator.py

+++ python_introduction/finance_calculator.py
@@ -1,5 +1,5 @@
-monthly_income = int(input("Enter your monthly income: "))
-total_monthly_expenses = int(input("Enter your total monthly expenses: "))
+monthly_income = float(input("Enter your monthly income: "))
+total_monthly_expenses = float(input("Enter your total monthly expenses: "))
 monthly_savings = monthly_income - total_monthly_expenses
 annual_rate = 0.05
 projected_savings = monthly_savings * 12 + (monthly_savings * 12 * 0.05)
